requirements.txt
numpy==1.19.5
torch==1.8.1+cu111
pillow==8.4.0
matplotlib==3.2.2
torchvision==0.9.1+cu111
tqdm==4.62.3
opencv-python==4.1.2.30
scipy==1.4.1